Revert changes from #405 and #408

+++ include/jwt-cpp/jwt.h
@@ -1626,7 +1626,7 @@ namespace jwt {
 					ec = error::signature_generation_error::signinit_failed;
 					return {};
 				}
-				if (!EVP_DigestUpdate(ctx.get(), data.data(), static_cast<unsigned int>(data.size()))) {
+				if (!EVP_DigestUpdate(ctx.get(), data.data(), data.size())) {
 					ec = error::signature_generation_error::digestupdate_failed;
 					return {};
 				}
@@ -1666,7 +1666,7 @@ namespace jwt {
 					ec = error::signature_verification_error::verifyinit_failed;
 					return;
 				}
-				if (!EVP_DigestUpdate(ctx.get(), data.data(), static_cast<unsigned int>(data.size()))) {
+				if (!EVP_DigestUpdate(ctx.get(), data.data(), data.size())) {
 					ec = error::signature_verification_error::verifyupdate_failed;
 					return;
 				}
@@ -1985,7 +1985,7 @@ namespace jwt {
 					return {};
 				}
 #endif
-				if (EVP_DigestUpdate(md_ctx.get(), data.data(), static_cast<unsigned int>(data.size())) != 1) {
+				if (EVP_DigestUpdate(md_ctx.get(), data.data(), data.size()) != 1) {
 					ec = error::signature_generation_error::digestupdate_failed;
 					return {};
 				}
@@ -2034,7 +2034,7 @@ namespace jwt {
 					return;
 				}
 #endif
-				if (EVP_DigestUpdate(md_ctx.get(), data.data(), static_cast<unsigned int>(data.size())) != 1) {
+				if (EVP_DigestUpdate(md_ctx.get(), data.data(), data.size()) != 1) {
 					ec = error::signature_verification_error::verifyupdate_failed;
 					return;
 				}
@@ -2511,14 +2511,14 @@ namespace jwt {
 		template<typename string_type, typename integer_type>
 		using is_substr_start_end_index_signature =
 			typename std::is_same<decltype(std::declval<string_type>().substr(
-									  static_cast<size_t>(std::declval<integer_type>()),
-									  static_cast<size_t>(std::declval<integer_type>()))),
+									  std::declval<integer_type>(),
+									  std::declval<integer_type>())),
 								  string_type>;
 
 		template<typename string_type, typename integer_type>
 		using is_substr_start_index_signature =
 			typename std::is_same<decltype(std::declval<string_type>().substr(
-									  static_cast<size_t>(std::declval<integer_type>()))),
+									  std::declval<integer_type>())),
 								  string_type>;
 
 		template<typename string_type>
@@ -2676,7 +2676,7 @@ namespace jwt {
 		date as_date() const {
 			using std::chrono::system_clock;
 			if (get_type() == json::type::number)
-				return date(std::chrono::seconds(static_cast<int64_t>(std::llround(as_number()))));
+				return date(std::chrono::seconds(std::llround(as_number())));
 			return date(std::chrono::seconds(as_integer()));
 		}
